Moved session setup to a middleware Co-authored-by: Josh Lakenan <joshualakenan@gmail.com> Co-authored-by: Alan Cho <alancho397@gmail.com> Co-authored-by: Trevor Rees-Jones <treesjones3@gmail.com>
