+++ src/api/init_api.js
@@ -1,7 +1,6 @@
 import express from "express";
 import dotenv from "dotenv";
 import pinoHttp from "pino-http";
-import fs from "fs";
 import { resolve } from "path";
 dotenv.config();
 
@@ -17,15 +16,7 @@ import errorHandler from "./middleware/error_handler.js";
 import sanitize from "./middleware/sanitize.js";
 import helmet from "helmet";
 import cors from "cors";
-
-//Session
-import session from "express-session";
-import store from "better-sqlite3-session-store";
-import sqlite from "better-sqlite3";
-
-const SqliteStore = store(session);
-if (!fs.existsSync("pnpd_data")) fs.mkdirSync("pnpd_data");
-const db = new sqlite("pnpd_data/session.db");
+import sessionConfig from "./middleware/session_config.js";
 
 function initApi(app) {
   const server = express();
@@ -39,25 +30,7 @@ function initApi(app) {
       credentials: true,
     })
   );
-  server.use(
-    session({
-      store: new SqliteStore({
-        client: db,
-        expired: {
-          clear: true,
-          intervalMs: 900000, //ms = 15min
-        },
-      }),
-      secret: process.env.SESSION_SECRET || "secret",
-      resave: false,
-      saveUninitialized: true,
-      cookie: {
-        maxAge: 60 * 60 * 1000, // 1 hour
-        httpOnly: true,
-        secure: false,
-      },
-    })
-  );
+  server.use(sessionConfig());
 
   server.use(pinoHttp({ stream: app.logger.sqliteStream() }));
   server.use(sanitize());
